detector/requirements.txt
torch==2.8.0
transformers==4.57.3
huggingface-hub==0.36.0
safetensors==0.7.0
pillow==11.3.0
numpy==2.0.2
requests==2.32.5
urllib3==1.26.6
